Updated the sample
