+++ Xamarin.Forms.Controls.Issues/Xamarin.Forms.Controls.Issues.Shared/Issue11831.xaml
@@ -1,0 +1,105 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<controls:TestContentPage
+    xmlns:controls="clr-namespace:Xamarin.Forms.Controls"
+    xmlns="http://xamarin.com/schemas/2014/forms"
+    xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+    xmlns:d="http://xamarin.com/schemas/2014/forms/design"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    mc:Ignorable="d"
+    x:Class="Xamarin.Forms.Controls.Issues.Issue11831"
+    x:Name="Issue11831Page"
+    Title="Issue 11831">
+    <controls:TestContentPage.Resources>
+        <ResourceDictionary>
+
+            <Style x:Key="CollectionSwipeSelectedStyle" TargetType="SwipeView">
+                <Setter Property="VisualStateManager.VisualStateGroups">
+                    <VisualStateGroupList>
+                       <VisualStateGroup x:Name="CommonStates">
+                           <VisualState x:Name="Normal" />
+                            <VisualState x:Name="Selected">
+                                <VisualState.Setters>
+                                    <Setter Property="BackgroundColor" Value="Green" />
+                                </VisualState.Setters>
+                            </VisualState>
+                        </VisualStateGroup>
+                    </VisualStateGroupList>
+                </Setter>
+            </Style>
+
+        </ResourceDictionary>
+    </controls:TestContentPage.Resources>
+    <Grid
+        RowSpacing="0">
+        <Grid.RowDefinitions>
+            <RowDefinition Height="Auto" />
+            <RowDefinition Height="*" />
+        </Grid.RowDefinitions>
+        <Label
+            Padding="12"
+            BackgroundColor="Black"
+            TextColor="White"
+            Text="Select am item from the list. If the selection background color is green, the test has passed."/>
+        <Grid
+            Grid.Row="1">
+            <CollectionView
+                x:Name="DataListView"
+                SelectionMode="Single"
+                ItemsSource="{Binding Data}">
+                <CollectionView.ItemTemplate>
+                    <DataTemplate x:DataType="x:String">
+                       <SwipeView
+                           BackgroundColor="Red"
+                           Style="{StaticResource CollectionSwipeSelectedStyle}">
+                            <SwipeView.RightItems>
+                                <SwipeItems>
+                                    <SwipeItemView
+                                        Command="{Binding BindingContext.DeleteCommand, Source={x:Reference Issue11831Page}}"
+                                        CommandParameter="{Binding}">
+                                        <Grid
+                                            BackgroundColor="LightSlateGray">
+                                            <Frame
+                                                BackgroundColor="#de3f45"
+                                                BorderColor="White"
+                                                Margin="8"
+                                                Padding="10"
+                                                CornerRadius="5">
+                                                <StackLayout
+                                                    Orientation="Vertical"
+                                                    VerticalOptions="Center"
+                                                    HorizontalOptions="Center">
+                                                    <Label
+                                                        Text="Delete"
+                                                        TextColor="White"
+                                                        FontSize="Medium"
+                                                        FontAttributes="Bold"
+                                                        LineBreakMode="NoWrap"
+                                                        WidthRequest="64"
+                                                        HorizontalTextAlignment="Center" />
+                                                </StackLayout>
+                                            </Frame>
+                                        </Grid>
+                                    </SwipeItemView>
+                                </SwipeItems>
+                            </SwipeView.RightItems>
+                            <Frame
+                                Padding="1"
+                                BorderColor="Silver"
+                                BackgroundColor="WhiteSmoke"
+                                HeightRequest="60"
+                                CornerRadius="12">
+                                <Label
+                                    Text="{Binding}"
+                                    TextColor="Black"
+                                    FontAttributes="Bold"
+                                    FontSize="Medium"
+                                    HorizontalTextAlignment="Center"
+                                    VerticalOptions="Center"/>
+                            </Frame>
+                        </SwipeView>
+                    </DataTemplate>
+                </CollectionView.ItemTemplate>
+            </CollectionView>
+        </Grid>
+    </Grid>
+</controls:TestContentPage>
